web-animations/animation-effect/animation-effect-iteration-composite.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffect firstChild attribute test</title>
<meta name="assert" content="The iteration composite operation property of this animation effect as specified by one of the IterationCompositeOperation enumeration values">
<link rel="help" href="https://web-animations.github.io/web-animations-spec/bikeshed/Overview.html#dom-animationeffect">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var effectsToTest = [newKeyframeEffect(), newMotionPathEffect()];
    effectsToTest.forEach(function(effect) {
        assert_equals(effect.iterationComposite, 'replace', 'Wrong default value of ' +
            'iterationComposite attribute of effect ' + type(effect));
    });
}, 'Test that iterationComposite attribute returns default value of iteration composite ' +
    'if it was not set by the constructor');

test(function() {
    var effectsToTest = [newKeyframeEffect({iterationComposite: 'accumulate'}),
        newMotionPathEffect({iterationComposite: 'accumulate'})];
    effectsToTest.forEach(function(effect) {
        assert_equals(effect.iterationComposite, 'accumulate', 'Wrong value of ' +
            'iterationComposite attribute of effect ' + type(effect));
    });
}, 'Test iterationComposite attribute returns value set by the constructor');

async_test(function(t) {
    var test = this;
    var effectsToTest = [newKeyframeEffect(), newMotionPathEffect()];
    effectsToTest.forEach(function (effect) {
        effect.iterationComposite = 'accumulate';
        // Now check that iteration composition is really 'accumulate'
        var target = createDiv(test);
        var animation = new Animation(target, effect,
            {iterations: 3, duration: ANIMATION_END_TIME});
        var player = document.timeline.play(animation);
        player.currentTime = ANIMATION_END_TIME - 10;

        setTimeout(function() {
            t.step(function () {
                // After the first iteration target element top == 200px
                // After 100 ms of the second iteration width should be
                // 200 + 'width at 100 ms'
                assert_approx_equals(window.getComputedStyle(target).top,
                    ANIMATION_TOP_1 + getExpectedTop(100), EPSILON,
                    'Iteration composition should be \'accumulate\'');
                // Now seek to the end, change iteration composition
                // to replace and check
                player.currentTime = 2 * ANIMATION_END_TIME - 10;
                effect.iterationComposite = 'replace';
                setTimeout(function() {
                    t.step(function () {
                        assert_approx_equals(window.getComputedStyle(target).top,
                            getExpectedTop(100), EPSILON,
                            'Iteration composition should be \'replace\'');
                        t.done();
                    });
                }, 110);

            });
        }, 110);        
    });
}, 'Test that setting iterationComposite attribute changes iteration composition ' +
    'according to the new value');
</script>
</body>
